added new code to extract video clips and tile them into arrays
these are similar to previous versions for Prey-Capture-SSM, but they are now generic so they should work with any set of video clips, all you need are source filenames and frame ranges. They should also be much more robust to video codec variations across machines.
ExtractVideoClip makes one clip at a time, and is intended to be run in a loop to dump a lot of clips into an output folder.
TileVidClips can then tile all the clips in that folder into a customizable nxn array